[refactor] Routes, sub-folders, ...
CREP calendar is now under the route `/crep`
CREP register form is now under the route `/crep/register`

For the moment, the route `/` returns a 404, this is intentional. The main page will come...
